Title: 학점 계산기, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 260112/학점 계산기/credit-calculator.cpp
@@ -13,7 +13,7 @@ int main() {
         sum_val += arr[i];
     }
 
-    double avg = (double)sum_val / n;
+    double avg = sum_val / n;
 
     cout << fixed;
     cout.precision(1);
